refactor(rules-injector): extract match reason resolver
